Move the import logic into the import package.
