+++ cmd/algorand-indexer/main.go
@@ -16,6 +16,14 @@ import (
 	"github.com/algorand/indexer/version"
 )
 
+func maybeFail(err error, errfmt string, params ...interface{}) {
+	if err == nil {
+		return
+	}
+	fmt.Fprintf(os.Stderr, errfmt, params...)
+	os.Exit(1)
+}
+
 var rootCmd = &cobra.Command{
 	Use:   "indexer",
 	Short: "Algorand Indexer",
@@ -106,7 +114,6 @@ func init() {
 		viper.AddConfigPath(k)
 	}
 	viper.SetEnvKeyReplacer(strings.NewReplacer("-", "_"))
-
 
 	if err := viper.ReadInConfig(); err != nil {
 		if _, ok := err.(viper.ConfigFileNotFoundError); ok {
